requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
pydeck>=0.8.0
plotly>=5.18.0
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0

# Per NBO Dashboard
folium>=0.15.0
streamlit-folium>=0.15.0
geopy>=2.4.0
tqdm>=4.66.0
